feat!: support private enums
If the enum name starts with a lowercase letter, then all related
symbols are also private.

Also change generated symbols that are meant to be internal to start
with `_` followed by a lower-case letter. These are private regardless
of the enum's visibility.

Some errors are now public (if the enum is) as they seemed likely
relevant to other packages.